backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
SQLAlchemy
newsapi-python
textblob
spacy
python-jose[cryptography]
passlib[bcrypt]
redis
slowapi
yfinance
authlib
httpx
python-multipart
cryptography
google-generativeai
beautifulsoup4
